free up resource.id from importing too much
