+++ src/inmanta/data/__init__.py
@@ -46,6 +46,7 @@ from asyncpg.exceptions import SerializationError
 from asyncpg.protocol import Record
 
 import inmanta.db.versions
+import inmanta.protocol
 import inmanta.types
 from crontab import CronTab
 from inmanta import const, resources, util
