+++ src/inmanta/export.py
@@ -28,12 +28,12 @@ import pydantic
 
 from inmanta import const, loader, protocol
 from inmanta.agent.handler import Commander
-from inmanta.ast import CompilerException, Namespace
+from inmanta.ast import CompilerException, Namespace, UnknownException
 from inmanta.ast.entity import Entity
 from inmanta.config import Option, is_list, is_uuid_opt
 from inmanta.const import ResourceState
 from inmanta.data.model import PipConfig
-from inmanta.execute.proxy import DynamicProxy, UnknownException
+from inmanta.execute.proxy import DynamicProxy
 from inmanta.execute.runtime import Instance
 from inmanta.module import Project
 from inmanta.protocol import Result
