+++ tests/compiler/test_null.py
@@ -17,7 +17,7 @@
 """
 
 import inmanta.compiler as compiler
-from inmanta.execute.proxy import UnsetException
+from inmanta.ast import UnsetException
 from inmanta.execute.util import NoneValue
