+++ tests/db/migration_tests/test_v202410310_to_v202411140.py
@@ -24,7 +24,6 @@ import asyncpg
 import pytest
 
 import inmanta.types
-from inmanta import resources
 from inmanta.data import Environment, ResourcePersistentState, Scheduler
 from inmanta.deploy import state
 from inmanta.deploy.state import ComplianceStatus
