+++ tests/test_docs_snippets.py
@@ -31,7 +31,6 @@ import py
 import pytest
 
 from inmanta import data
-from inmanta.data import model
 from utils import v1_module_from_template, wait_until_deployment_finishes
 
 DOCS_DIR: str = os.path.join(os.path.dirname(__file__), "..", "docs")
